pyproject.toml
[project]
name = "knowledger"
version = "0.1.0"
description = "An intelligent knowledge management system with Telegram bot and web interface"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.124.4",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
    "python-telegram-bot>=22.5",
    "openai>=1.0.0",
    "pyvis>=0.3.2",
    "networkx>=3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
]
